hotfix for appveyor

+++ test/tensor/test_algorithms.cpp
@@ -312,9 +312,9 @@ BOOST_FIXTURE_TEST_CASE_TEMPLATE( test_tensor_algorithms_trans, value,  test_typ
 		for(auto i = 0ul; i < s; ++i)
 			BOOST_CHECK_EQUAL( a[i], b2[i] );
 
-		long unsigned int zero = 0;
+		size_t zero = 0;
                 ublas::trans( zero, n.data(), pi.data(), c2.data(), wc.data(), a.data(), wa.data() );
-                ublas::trans( zero, nc.data(), pi.data(), b2.data(), wb.data(),    c2.data(), wc.data() );
+                ublas::trans( zero, nc.data(), pi.data(), b2.data(), wb.data(), c2.data(), wc.data() );
 
                 value_type *c0 = nullptr;
                 size_type const*const s0 = nullptr;
